Update README.md
Today, the area of point-of-care diagnostics is synonymous with paper microfluidics where cheap, disposable, and on-the-spot detection toolkits are being developed for a variety of chemical tests. In this work, we present a novel application of microfluidic paper-based analytical devices (μPADs) to study the behavior of a small model nematode, Caenorhabditis elegans. A custom program is written to automatically recognize individual worms on the μPADs and extract locomotion parameters in real-time.